#86bw57fuw Changed position data to Double

+++ src/main/scala/org/aas/sbtanks/entities/tank/behaviours/TankMovementBehaviour.scala
@@ -1,11 +1,11 @@
 package org.aas.sbtanks.entities.tank.behaviours
 
-trait TankMovementBehaviour(startingX: Int = 0, startingY: Int = 0):
+trait TankMovementBehaviour(startingX: Double = 0, startingY: Double = 0):
     private var position = (startingX, startingY)
 
-    def positionX: Int = position(0)
+    def positionX: Double = position(0)
 
-    def positionY: Int = position(1)
+    def positionY: Double = position(1)
 
-    def move(amountX: Int, amountY: Int) = 
+    def move(amountX: Double, amountY: Double) = 
         position = (position(0) + amountX, position(1) + amountY)
